Rust style fn eal_init is defined.
